Fixed Minor Bug from basket fragment. Added ZXING Android library and its code for scanning QR code.

+++ app/src/main/java/com/roohia/hp/laundry/gui/Fragments/BasketFragment.java
@@ -89,7 +89,7 @@ public class BasketFragment extends Fragment implements  AdapterView.OnItemClick
         switch (view.getId()){
             case R.id.btn_remove:
                 if(basketItemsAdapter != null) {
-                    basketItemsAdapter.getItems().remove(view.getTag());
+                    basketItemsAdapter.getItems().remove((int)view.getTag());
                     basketItemsAdapter.notifyDataSetChanged();
                 }
                 break;
